// loja-kgv/src/data/Types.ts
export type TypeCarro = {
    id: number;
    nome: string;
    foto1: string;
    foto2: string;
    foto3: string;
    foto4: string;
    foto5: string;
    marca: string;
    modelo: string;
    versao: string;
    preco: string;
    anoFabricacao: number;
    anoModelo: number;
    km: string;
    motor: string;
    potencia: string;
    aceleracao: string;
    airbag: number;
    combustivel: string;
    conservacao: string;
    transmissao: string;
    placaFim: number;
    portas: number;
    cor: string;
    bancos: string;
    freios: string;
    farois: string;
    rodas: string;
    blindado: string;
    bluetooth: string;
    tetoSolar: string;
    conversivel: string;
    pilotoAutomatico: string;
    multimidia: string;
    sensorChuva: string;
    sensorEstacionamento: string;
  };